posts_comments_api/requirements.txt
Django
django-cors-headers
django-ninja
django-ninja-jwt[crypto]
pydantic
requests
django-dotenv
celery